docs_requirements.txt
sphinx>=5.3.0
furo>=2022.9.29
sphinx_design>=0.3.0
sphinx_copybutton>=0.5.1
setuptools==59.8.0
click-man==0.4.1
